Base de datos Listas Proyecto Configurado

+++ Practica Final/Entidades/Juegos.cs
@@ -7,12 +7,12 @@ namespace Practica_Final.Entidades
 {
    public class Juegos
     {
+       
         [Key]
         public int JuegoId { get; set; }
         public DateTime FechaCompra { get; set; } = DateTime.Now;
         public string Descripcion { get; set; }
-        public int Precio { get; set; }
+        public double Precio { get; set; }
         public int Existencia { get; set; }
-
     }
 }
